re #866 updating interson sandbox, unverified if workingMerged revision(s) 3882-4393 from trunk/PlusApp:
Merged revision(s)  from trunk/PlusLib:

Merged revision(s) 3882-4393 from trunk/PlusBuild:

+++ branches/rankin-sandbox/PlusLib/src/Utilities/MicronTrackerInterface/Facet.cpp
@@ -118,7 +118,7 @@ bool Facet::identify(MCamera* cam, std::vector<Vector*> vectorSet, double positi
   else
     camHandle = cam->Handle();
 
-  Facet_Identify(this->m_handle, cam->Handle(), vectorHandles, vectorSet.size(),  &result);
+  Facet_Identify(this->m_handle, cam->Handle(), vectorHandles, static_cast<int>(vectorSet.size()),  &result);
   free(vectorHandles);
   return result;
 }
